// src/app/shared/enums/displayValues.enum.ts
export enum DisplayValues {
  // digits
  Digit0 = '0',
  Digit1 = '1',
  Digit2 = '2',
  Digit3 = '3',
  Digit4 = '4',
  Digit5 = '5',
  Digit6 = '6',
  Digit7 = '7',
  Digit8 = '8',
  Digit9 = '9',

  // operations
  Add = '+',
  Subtract = '-',
  Multiply = '*',
  Divide = '/',
  Equals = '=',

  // percent and point
  Percent = '%',
  Point = '.',

  // functions
  AllClear = 'AC',
  Clear = 'C',
  ToggleTheme = '🌓',
}