// src/app/shared/enums/operations.enum.ts
export enum Operations {
  // digits
  Digit0 = 'digit0',
  Digit1 = 'digit1',
  Digit2 = 'digit2',
  Digit3 = 'digit3',
  Digit4 = 'digit4',
  Digit5 = 'digit5',
  Digit6 = 'digit6',
  Digit7 = 'digit7',
  Digit8 = 'digit8',
  Digit9 = 'digit9',

  // operations
  Add = 'add',
  Subtract = 'subtract',
  Multiply = 'multiply',
  Divide = 'divide',
  Equals = 'equals',

  // percent and point
  Percent = 'percent',
  Point = 'point',

  // functions
  AllClear = 'all-clear',
  Clear = 'clear',
  ToggleTheme = 'toggle-theme',
}